+++ plone/app/layout/viewlets/social.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from plone import api
 from plone.app.layout.viewlets.common import TitleViewlet
 from plone.memoize.view import memoize
 from plone.registry.interfaces import IRegistry
@@ -41,6 +42,9 @@ class SocialTagsViewlet(TitleViewlet):
                                          check=False)
 
         if not settings.share_social_data:
+            return []
+
+        if not api.user.is_anonymous():
             return []
 
         tags = [